feat: modulariza exibição de livros e skeletons na busca e página inicial com criação do arquivo renderBooks
